# save log waste and tag caching