@ss920386 2024-05-06
Thanks for the PR. May I ask will this have a risk of memory leak? I didn't see sync.Map.Delete() called in this PR. Will the cache grow forever?