@suhuaqin 2024-05-06
> @suhuaqin Thanks for the PR. May I ask will this have a risk of memory leak? I didn't see sync.Map.Delete() called in this PR. Will the cache grow forever?

The maximum size of the cache depends on the number of structures to be parsed and the number of members of the structure. If you are concerned about memory wastage caused by the small frequency of use of certain caches, consider introducing a cache elimination mechanism